moved /examples/kuka_iiwa_arm/dev/tools/simple_tree_visualizer to /manipulation/util/simple_tree_visualizer
